// apps/cowswap-frontend/src/modules/priceChart/lib/priceChartAdvancedApi.types.ts
export interface Bar {
  close: number
  high: number
  low: number
  open: number
  time: number
  volume?: number
}

export interface ChartingLibraryWidgetConstructor {
  new (options: ChartingLibraryWidgetOptions): IChartingLibraryWidget
}

export type ChartPropertiesOverrides = Record<string, string | number | boolean>

export interface CustomFormatters {
  priceFormatterFactory?: () => {
    format(value: number): string
  }
}

export interface DatafeedConfiguration {
  exchanges: {
    desc: string
    name: string
    value: string
  }[]
  supported_resolutions: ResolutionString[]
  supports_time: boolean
}
export interface IBasicDataFeed {
  getBars(
    symbolInfo: LibrarySymbolInfo,
    resolution: ResolutionString,
    periodParams: PeriodParams,
    onResult: HistoryCallback,
    onError: ErrorCallback,
  ): void
  onReady(callback: OnReadyCallback): void
  resolveSymbol(symbolName: string, onResolve: (symbol: LibrarySymbolInfo) => void, onError: ErrorCallback): void
  searchSymbols(userInput: string, exchange: string, symbolType: string, onResult: SearchSymbolsCallback): void
  subscribeBars(
    symbolInfo: LibrarySymbolInfo,
    resolution: ResolutionString,
    onRealtimeCallback: (bar: Bar) => void,
    subscriberUid: string,
    onResetCacheNeededCallback: () => void,
  ): void
  unsubscribeBars(subscriberUid: string): void
}

export interface IChartingLibraryWidget {
  activeChart(): ChartApi
  applyOverrides(overrides: Partial<ChartPropertiesOverrides>): void
  changeTheme(theme: 'dark' | 'light'): Promise<void>
  onChartReady(callback: () => void): void
  remove(): void
  save(callback: (state: object) => void): void
  subscribe(event: 'onAutoSaveNeeded', callback: () => void): void
  unsubscribe(event: 'onAutoSaveNeeded', callback: () => void): void
}

export interface LibrarySymbolInfo {
  data_status: string
  description: string
  exchange: string
  format: string
  has_daily: boolean
  has_intraday: boolean
  has_weekly_and_monthly: boolean
  listed_exchange: string
  minmov: number
  name: string
  pricescale: number
  session: string
  supported_resolutions: ResolutionString[]
  ticker?: string
  timezone: string
  type: string
  visible_plots_set: string
  volume_precision: number
}

export type OnReadyCallback = (configuration: DatafeedConfiguration) => void
export type ResolutionString = string

export interface SearchSymbolResultItem {
  description: string
  exchange: string
  symbol: string
  ticker?: string
  type: string
}

export type SearchSymbolsCallback = (items: SearchSymbolResultItem[]) => void

interface ChartApi {
  createStudy(name: string, forceOverlay: boolean, lock: boolean): Promise<string>
  getStudyById(id: string): { mergeUp(): void }
  getAllStudies(): { id: string; name: string }[]
  removeEntity(id: string, options?: { disableUndo?: boolean }): void
  setSymbol(symbol: string, callback: () => void): void
  symbol(): string
}

interface ChartingLibraryWidgetOptions {
  auto_save_delay?: number
  autosize: boolean
  container: string
  custom_css_url?: string
  custom_formatters?: CustomFormatters
  datafeed: IBasicDataFeed
  disabled_features?: string[]
  enabled_features?: string[]
  favorites?: {
    chartTypes: string[]
    intervals: ResolutionString[]
  }
  interval: ResolutionString
  library_path: string
  loading_screen?: {
    backgroundColor: string
    foregroundColor: string
  }
  locale?: string
  overrides?: Partial<ChartPropertiesOverrides>
  saved_data?: object
  symbol: string
  theme?: 'dark' | 'light'
  time_frames?: TimeFrame[]
  timezone?: string
}

type ErrorCallback = (reason: string) => void

type HistoryCallback = (bars: Bar[], metadata: { nextTime?: number; noData: boolean }) => void

interface PeriodParams {
  countBack: number
  firstDataRequest: boolean
  from: number
  to: number
}

interface TimeFrame {
  title?: string
  description: string
  resolution: ResolutionString
  text: string
}